refactor for the usage of winston logger
