+++ src/auth-service/middleware/passport.js
@@ -6,7 +6,8 @@ const Validator = require("validator");
 const UserSchema = require("@models/User");
 const AccessTokenSchema = require("@models/AccessToken");
 const constants = require("@config/constants");
-const { logElement, logText, logObject, winstonLogger } = require("@utils/log");
+const winstonLogger = require("@utils/log-winston");
+const { logElement, logText, logObject } = require("@utils/log");
 const { Strategy: JwtStrategy, ExtractJwt } = require("passport-jwt");
 const AuthTokenStrategy = require("passport-auth-token");
 const jwt = require("jsonwebtoken");
