+++ src/auth-service/utils/control-access.js
@@ -11,7 +11,8 @@ const httpStatus = require("http-status");
 const mongoose = require("mongoose").set("debug", true);
 const accessCodeGenerator = require("generate-password");
 const { getModelByTenant } = require("@config/database");
-const { logObject, logElement, logText, winstonLogger } = require("@utils/log");
+const winstonLogger = require("@utils/log-winston");
+const { logObject, logElement, logText } = require("@utils/log");
 const mailer = require("@utils/mailer");
 const generateFilter = require("@utils/generate-filter");
 const isEmpty = require("is-empty");
