+++ src/auth-service/utils/create-favorite.js
@@ -3,7 +3,7 @@ const UserSchema = require("@models/User");
 const httpStatus = require("http-status");
 const mongoose = require("mongoose").set("debug", true);
 const { getModelByTenant } = require("@config/database");
-const { logObject, logElement, logText, winstonLogger } = require("@utils/log");
+const { logObject, logElement, logText } = require("@utils/log");
 const generateFilter = require("@utils/generate-filter");
 const isEmpty = require("is-empty");
 const constants = require("@config/constants");
