+++ src/auth-service/utils/test/ut_log.js
@@ -1,15 +1,7 @@
 require("module-alias/register");
 const { expect } = require("chai");
 const sinon = require("sinon");
-const winston = require("winston");
-const MongoDB = require("winston-mongodb").MongoDB;
-const {
-  logText,
-  logElement,
-  logObject,
-  logError,
-  winstonLogger,
-} = require("@utils/log");
+const { logText, logElement, logObject, logError } = require("@utils/log");
 
 describe("logger", () => {
   describe("logText", () => {
@@ -116,43 +108,5 @@ describe("logger", () => {
       expect(result).to.equal("log deactivated in prod and stage");
       process.env.NODE_ENV = "test"; // Reset NODE_ENV to 'test'
     });
-  });
-
-  describe("winstonLogger", () => {
-    it("should have the correct log level", () => {
-      expect(winstonLogger.level).to.equal("info");
-    });
-
-    it("should have a MongoDB transport", () => {
-      expect(winstonLogger.transports).to.be.an("array");
-      expect(winstonLogger.transports).to.have.lengthOf(1);
-
-      const transport = winstonLogger.transports[0];
-      expect(transport).to.be.an.instanceOf(MongoDB);
-    });
-
-    it("should have the correct MongoDB transport options", () => {
-      const transport = winstonLogger.transports[0];
-
-      expect(transport.db).to.be.a("string"); // Replace with the correct value based on your LogDB function
-      expect(transport.options).to.deep.equal({
-        useNewUrlParser: true,
-        useUnifiedTopology: true,
-      });
-      expect(transport.collection).to.equal("logs");
-      expect(transport.options).to.deep.equal({
-        useUnifiedTopology: true,
-      });
-      expect(transport.format).to.be.an("object");
-      expect(transport.format.timestamp).to.be.a("function");
-      expect(transport.format.json).to.be.a("function");
-      expect(transport.format.metadata).to.be.a("function");
-      expect(transport.metaKey).to.equal("metadata");
-      expect(transport.level).to.equal("info");
-      expect(transport.schema).to.be.an("object"); // Replace with the correct value based on your LogSchema
-      expect(transport.model).to.be.an("object"); // Replace with the correct value based on your LogModel
-    });
-
-    // You can add more tests for specific functionalities of the winstonLogger if needed
   });
 });
